+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.2.0
 
 - Add in ability to download worlds from CurseForge
+- Fix mods from curse able to be installed with no loader